[FIX] documents: fix documents creation on documents
1. Fix reply with an attachment on message from existing document

How to reproduce:
- Go to the chatter of a document
- Add yourself as a follower
- Send a message from the chatter: you should receive an email
- Reply to this email with an attachment

You receive a delivery error message ("Address not found") and your reply
doesn't appear in the chatter, nor your attachment.
With this fix, the message with its attachment appears in the chatter and there
is no delivery error message.

2. Fix sending a message with an attachment on a request

How to reproduce:
- Go to the chatter of a request
- Send a message with an attachment from its chatter

You get the error "The operation cannot be completed: This attachment is
already a document." while the message with its attachment should be added in
the chatter.
With this fix, the message with its attachment appears in the chatter and the
request remains to be fulfilled.

3. Error while uploading a requested document through activity popover

How to reproduce:
- As user A, create a requested document for User B
- As user B, click on the activity of that request and upload a document

You get the error "The operation cannot be completed: This attachment is
already a document." while the uploaded file should be linked to the request.
With this fix, the request is fulfilled, and you get no error.

We solve those problems by not creating document by default when messages with
attachments are posted on a document and delegate the creation of the documents
to the _message_post_after_hook method for the case of a post on a folder
alias. We have chosen that default behavior as in all uses case except the
alias, we expect the attachment to be added to the chatter and not as a new
document.

TestMailGateway already tests that incoming mail with attachment on an alias
create the related documents, and we introduce a test for each case described
above in other commits.

Note that we also simplify the API by always ignoring the "no_document"
environment variable in _message_post_after_hook as it is always set to true in
message_post.

Task-3707821

X-original-commit: 05311d022cc60045afafd97622913398867e022d
Part-of: odoo/enterprise#80289
Related: odoo/odoo#199479
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Pierre-Yves Dufays (pydu) <pydu@odoo.com>