[PERF] approvals_purchase: improve performance
To log the changes in the state of a purchase order to the chatter of the related approval requests,
a search was done to find the product lines related to the approval requests.  This search was
performed inside a loop.

This commit performs the search outside of the loop to avoid slowing down the system when the
state of multiple purchase orders has changed.

Related-PR: https://github.com/odoo/enterprise/pull/71294

task-3561514

closes odoo/enterprise#79559

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>