[FIX] account_accountant: missing colspan on lock date wizard
The items inside the wizard view are ordered as a 2 column grid of
labels and contents. But recently, it is found that the hard lock date
alert item is only displayed in the short label section of the grid.
This commit spans the hard lock date `div`'s `colspan` to 2 to make sure
it spans on the whole line.

task-4586675

closes odoo/enterprise#79616

X-original-commit: a2412f3d619f54b18f0899db57cf3f5f1cf3ffc4
Signed-off-by: William André (wan) <wan@odoo.com>