[FIX] documents: fix 'space' selection in Kanban
This commit fixes the selection in Kanban. When refactoring the selection,
it was intended to allow selecting records in Documents, but the feature
was not working as expected and it was only available when the selection
was already active.

Now, it has been decided to make this feature standard for Kanban views,
which allows us to remove this override.

task-4582340

X-original-commit: ca8a80047ad9b75dd93e9423fa9da9642f8b9527
Part-of: odoo/enterprise#79752
Related: odoo/odoo#198422
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>
Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>