[IMP] web_studio: allow header button edition in Kanban
This commit introduce for the Kanban editor a feature already available
in List view. As the feature behavior should be identical, the same
component has been made to make sure we can continue evolving the feature
for both views more conveniently.

Invisible buttons can now be edited properly in the Editor, while being
hidden by default, and having the same appearance as other invisible
elements while 'Show Invisible Elements' is checked.

The commit also moves the ButtonProperties component, as it is
now used in different editors, and not just in Form view.

Tests and tours have been modified, to mock the new route, and adapt to
the new way to edit attributes on the button node without having a
dedicated modal. Some selectors have also been changed.

task-4318325

a

Part-of: odoo/enterprise#80198
Related: odoo/odoo#199373
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>