[FIX] knowledge: update knowledge_commands_tour to `div` in composer
`test_knowledge_commands_tour` is currently disabled. Currently it is failing
because the main paragraph element in the Composer was changed from `<p>` to
`<div class="o-paragraph">`. This commit updates the tour to be compatible with
the changes in the Composer, in hope to re-enable the test some day.

runbot-task-114943
runbot-task-134158

closes odoo/enterprise#79952

X-original-commit: 5d205b8af8b53369f7be09a82f67b8e425de16b0
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>