[FIX] web_studio: all fields available for expressions (modifiers)
Since commit odoo/odoo@6f06420e4a9443c52dc0cb427f8f55eb4aecabce, fields used in expressions in a view
are automatically added to the view if not present originally.

This allows all fields to be available in the expression editor for modifiers, which this commit is enabling.

task-4481934

closes odoo/enterprise#79817

X-original-commit: 58557b5211cabab68016a75e5fd16bb92fef1288
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>