[FIX] website_knowledge: provide html for crawler on public articles
Ensure that the HTML for Knowledge published articles is not delayed by OWL
rendering promise.

The content of the article should be available for web crawlers even with
JavaScript disabled.

The sidebar will still be loaded as a Component for convenience, and
PublicWidget introduced through this commit will be converted to Interaction in
future versions.

task-4563216

closes odoo/enterprise#80963

X-original-commit: b8ca55ee1e68ce0eff62b739ac0ce2e09c9cf375
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>