[IMP] pos_restaurant_appointment: Add minimal html editor in pos mode
In the point of sale environnement we want to be able to edit html field
in a minimal way. So only core plugins are added to the html_editor.

As we have a special assets environnement for the point_of_sale we can
directly patch the main `htmlField` definition without impacting main
assets of Odoo.

closes odoo/enterprise#81717

Community: 198482
X-original-commit: 19110e430dc2501a20e583f51a6f6b428f2b51c1
Related: odoo/odoo#202392
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: David Monnom (moda) <moda@odoo.com>