[IMP] Batch workflow refactor
Problem
----------
Change the UX of payroll batches.

Objective
----------
Batches are now Pay Runs.
No more form view in full size.
Pay Run information is displayed above the list of payslips when they are filtered by a specific pay run.
All pay runs are accessible from a vertical kanban view.
To create a new kanban, there is a wizard in 2 steps.
One to select the dates, structure_id, etc. of the pay run
And the other to select the employees that fit the constraints

Solution
----------
All action buttons, smart buttons and fields of the form are used to display the pay run card.
The form arch is fetched and compiled in the kanban view and list view of payslips.
The wizard of pay run creation is the form view of payrun.
The wizard to select employees is the employee list view.

task-4568268

closes odoo/enterprise#87479

Related: odoo/odoo#213918
Related: odoo/upgrade#7861
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>