[FIX] account_batch_payment: create a batch with branch payments
Users managing several branch companies may collect in a batch several
payments from the branches. However this currently does not work as
expected.

Steps to reproduce:
- Have a [main] company and a [branch] company
- Create a payment in [branch]
- Create a batch in [main]
- Add the payment to the batch

Issue: Error will raise

```
Incompatible companies on records:
- “PAY00001” belongs to company “Your branch” and “Batch Payment”
  (batch_payment_id: 'BATCH/IN/2025/0001 (New)') belongs to another
  company
```

This occurs because now we check that batch and payment belong to the
same company
It is a change of behavior as in older versions this flows was
permitted

opw-4574188

closes odoo/enterprise#79959

X-original-commit: cc1660da890887d2057b3d7d129ab6ece90d1862
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Andrea Grazioso (agr) <agr@odoo.com>