[REF] documents: add basic documents view hoot tests
Also impacted: documents_spreadsheet tests. Note that we kept the
refactoring regarding res.users introduced in the original PR even
though it is not necessary to have an odoobot user now, but it will
allow to not do this work again if other extra setup becomes needed
for Documents.

Task-3861500

X-original-commit: b5cc3c15bbe96fe7507eab47d1b08e38aae14409
Part-of: odoo/enterprise#81976
Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>