[FIX] web_studio: display empty fields in (new) kanban editor
The kanban has been reworked with https://github.com/odoo/odoo/pull/167751
and odoo/enterprise@23426182d92198cf977d7eecbfb3d47117d80c59

Before this commit when a field is empty (the record has no value for a field) in the
kanban editor -- new version -- that field had no size and was virtually invisible.

This was because one little line of code was missing to enable the feature of applying a placeholder
in the empty field to make it visible.

opw-4526492

closes odoo/enterprise#80187

X-original-commit: 8e41255c0bbbf4c9a496ecdbefa81b4108051515
Signed-off-by: Bruno Boi (boi) <boi@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>