[FIX] stock_barcode: fix incrementing quantity from digipad view
This commit fixes incrementing quantities using packaging button in
digipad view. Previously, incrementing product quantity from packaging
button always increased the quantity in the factor of the packaging
unit, whatever the base unit of the product was. Therefore, it was
logically correct only if the product base unit had a factor of 1. After
this commit, the factor of the product base unit is taken into
consideration.

Task-4471379

Part-of: odoo/enterprise#84676
Related: odoo/odoo#208431
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Mohammad Abdulmoneim (abdu) <abdu@odoo.com>