[FIX] documents(,_hr_payroll): fix inherited and default members
+ tests update in documents_approval and _spreadsheet
+ query count increment in test_mail_enterprise

Fix owner edit role member after copy or creating inside a
folder, including while specifying access_ids.

Allow client code to add members to the default members added by
the base logic of documents (owner's log) and folders inherited
values => Fix HR Payroll Employee Declaration where members of the
payroll folder would not have access to the generated documents.

The paradigm is:
* The owner should always have a "document.access" record created
* It should be possible to add a member in addition to the inherited
  ones.
* If one passes values for a member when creating, it should override
  the default value for that partner.
* If access_ids is passed as `False`, it should prevent inheriting
  access (but not prevent a role=False record for owner).
* Follow-up of odoo/enterprise#77783, there is no need to enforce
  logic for owner's `documents.access` role as the owner is
  basically an empowered editor.
* On copying a document, the current user is the new document owner,
  except when copying a company root document

Additionally, make sure that
 * No membership (document.access with role) is created for odoobot
 * The access of owner is not shown on the permission panel

Note for fw ports:
As the current user is now the default owner, we better detect
attempts to create documents as superuser which triggers a
warning on runbots. Most of the logging could be muted in tests but
not when `project_sale_demo.xml` executes `action_confirm` which
creates documents.
To not delay solving the critical issue fixed, we switch the logging
level from warning to info, which is also visible by default.

Task-4593290

Added in the forward port (18.1+): As we now take current user as default
owner when duplicating documents, our `"Creating document(s) as superuser"`
was triggered in:
* tests, so we can simply mute it
* when copying a project with documents when documents_project_sale
  installed project_sale_demo.xml, at sale order confirmation, because we
  forgot to propagate the new owner to children when copying a folder.
  This is fixed here and will be backported.

closes odoo/enterprise#81004

X-original-commit: 649ee2e3b4a69842371698b69d4b767d387b2c68
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>