[FIX] hr_contract_salary: restore working schedule dropdown position
This commit restores the position of the working schedule
dropdown to be just below the label in the salary configurator,
like in previous versions.

The cause of the issue is that the SelectMenuWrapper component
is attached just after the invisible select element:
https://github.com/odoo/enterprise/blob/9f6a4f0697edcc9330ad7976c6602d892e7b6ef9/hr_contract_salary/static/src/js/hr_contract_salary.js#L69-L71
But before other elements are placed in the DOM between them.

A solution is to add a wrapper div around the selection, ensuring
the `SelectMenuWrapper` component is attached inside that div,
and further elements placed in the DOM after the wrapper div.

**Steps to reproduce**
- Install hr_contract_salary_payroll
- Select any employee contract > Generate offer > Salary configurator
- Add `part?=80` (any value != 100) at the end of URL

opw-4496279

closes odoo/enterprise#81108

X-original-commit: 45fa9805d5f4b0d56f40b3128e6471a06cc92e8d
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>