[FIX] account_external_tax: keep zero tax lines for external calculators
Tax lines are not created in cases where Odoo expects them to be $0
[1]. This is problematic for cases where tax is computed externally
and the tax is not in fact $0. There is now no tax line to adjust to
the right amount.

To reproduce:

1. Configure Avatax
2. Configure the Discount product to have the "Non-taxable product"
   Avatax category
2. Create an invoice to Azure Interior
3. Select the Avatax fiscal position
3. Add 1 FURN_6667 with default unit price $295
4. Add 1 Discount with unit price -$295
5. Click "Compute Taxes"

The invoice is fully discounted but $30.24 tax must still be
paid. This is correct. It's reflected in the invoice total, but the
tax lines are not in the journal items.

`_set_external_taxes()` goes through the summary to set the right tax
amounts, but cannot find the line.

This solves the issue by using the new key "__keep_zero_line" grouping
key in accounting that prevents these $0 tax lines from being
excluded.

opw-4493544

closes odoo/enterprise#81725

X-original-commit: 169a0679a1d76500158654820c3d2a502a799588
Related: odoo/odoo#202431
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>