[FIX] hr_contract_salary: partner mobile does not exist
Trying to create a partner with a value for 'mobile' which has been
recently removed.

By the way add mailing tests when having an offer linked to an applicant
without partner. It is not standard as partner are created when setting
email or phone, but it could happen. It is not completely well supported
currently.

Task-4555506: Cleanup / Mergeup default / suggested recipients

X-original-commit: 3802ba8928bf117fbfce20b28af48706050eae06
Part-of: odoo/enterprise#79999
Related: odoo/odoo#198936
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>