[FIX] approvals_purchase: properly override write for POs
The `write` method for `purchase.order` was not correctly overridden in
`approvals_purchase`. This commit fixes it so as to avoid unnecessary db
queries (i.e. `search` calls) and incorrect reassignment of the values
of the `self` recordset.

Also does a small improvement of using a defaultdict instead of manually
recreating its logic for a normal dict.

follow-up to:
odoo/enterprise#71294
odoo/enterprise#79559

closes odoo/enterprise#79953

Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>