[IMP] hr_referral: avoid generate a bunch of CSS rules for one animation
In hr_referral we have a special "progress bar" in a circular form.
To produce this effect, in SCSS, we have a loop to generate all states
(of the progress's percentage). But these loops generate a lot of rules
in CSS and so have an impact on the CSS bundle.

The CSS has improved a lot these last year, and now we can do amazing
thing in pure CSS so this commit aims to use new technics and avoid a
lot of glue code (for animation) and produce a similar result.

This commit use:
* CSS custom-property
* `conic-gradient`: to produce a pie (circular "progress bar")

also we scope SCSS variable to avoid side effect in other SCSS file.

closes odoo/enterprise#79954

Signed-off-by: Pierre Paridans (app) <app@odoo.com>