[FIX] {website_}appointment_crm: avoid duplicate tests
Until 18.0 [1], test classes should not inherit from other test classes which themselves contain tests.
When this happens, the tests of the parent class are run for every class that inherits from it.

We fix occurrences of this for social marketing apps.
A naive detection script is available on the pad of the related task.

[1]: 6dc96811c24ec4c97b8fbd2489aef6d4f061ac03

task-3792478

closes odoo/enterprise#80719

X-original-commit: d85fcc82927e5415bc8c94735f6fa43e1353acec
Related: odoo/odoo#200271
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Renaud Thiry (reth) <reth@odoo.com>