[IMP] web_gantt: focus date at the start of a group
This commit updates the behavior of focusDate to ensure that scrolling
stops when the start of the group containing the date is visible on the
left side of the grid, rather than centering the focused date within the
grid. Pressing the 'Today' button now follows the same logic but focuses
on the current date instead of the group.

Part of task-4571228

closes odoo/enterprise#80194

X-original-commit: e3623bc6bede66f070547d23c2546ed88f4c98e6
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>
Signed-off-by: Julien Carion (juca) <juca@odoo.com>