[MOV] l10n_br_test_avatax_sale: move tests to br_avatax_sale module
The l10n_br_test_avatax_sale module was kept after the refactor of
external_tax modules seen here because the code of the module was no
longer necessary but the test cases still had use.

Later down the line, the l10n_br_avatax_sale module was created to
account for CNAE, operation types, and other new features.

As this module expands it makes sense to move the test code back to
it instead of keeping a separate test module.

Part-of: odoo/enterprise#81525
Related: odoo/upgrade#7390
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>