[IMP] documents: improve the documents control view
Purpose
=======
Allow to show 2 columns when screen is 1280px wide.

Show only one embedded action, so it never displays on 2 rows.

Show the `i` button that open the chatter after the button to switch
view. Show an `x` instead of `i` when the chatter is opened.

When the details panel / chatter is opened, do not show the pager (and
show 80 records instead of 40 like before).

Task-4596337

Part-of: odoo/enterprise#80138
Related: odoo/odoo#201267
Related: odoo/upgrade#7475
Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>