[IMP] account_followup: Print is not ticked by default
Problem
---------
Follow-ups can be configured to execute multiple type of follow-ups
(email, whatsapp, sms, ...)

One of those is to print a letter, but it's not the most used. When you
try to execute a follow-up manually, a wizard opens and follows the
setup of the to-be-executed level ; so for example if the next level is
configured to send a SMS, SMS should be ticket by default.

But the "Print" option is always ticked, even if the "Letter" option is
not selected in the configuration of the level. This always generates an
additional PDF that is often not needed.

Objective
---------
Do not tick "Print" by default when manually executing a follow-up
level, unless "Letter" is ticket in the configuration of the level.

Solution
---------
Set "Print" default value in the wizard to the "Letter" value of the
follow up line instead of leaving it as True (which is the default).

task-4574971

fw port of: https://github.com/odoo/enterprise/pull/80112 and https://github.com/odoo/enterprise/pull/79829

closes odoo/enterprise#87868

X-original-commit: fd053735a65e7c31845dbce875cf09a61c8171a5
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>