[FIX] sign: clipboard tour fix
Before this commit, the `shared_sign_request_tour` tour was failing
in the clipboard check because the step just before clicked the toggle
making it hidden, which was not necessary as it is already shared.

After the commit, we no longer click the toggle to share the request
as it already opens as shared. Additionally, we propagate the download
button restriction from odoo/enterprise#80070.

runbot-116600

closes odoo/enterprise#80225

X-original-commit: 124714e6340c6ec37dc6f98a48164c0fbf20a200
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Gabriel de Paula Felix (gdpf) <gdpf@odoo.com>