[FIX] project_enterprise: fix traceback when use clicks get critical path
Currently, a traceback occurs when the user clicks critical path in tasks gantt.

To reproduce this issue:

1) Install Project and add the `Use Task Dependencies` group to the current user
2) Open the list view of the project
3) Enable the `Task Dependencies` to any one project
4) Now remove all the start dates from the Planned Date for all tasks
5) Open the Gantt view of the tasks and add a task
6) Refresh the page, and click the `Get Critical Path` button

Error:- "KeyError: 24"

https://github.com/odoo/enterprise/blob/c9a753fce85dfb09dc413831993e63c7736d258a/project_enterprise/models/project_task.py#L1544

From the above line, the traceback is occurring because the `path_last_task`
will take the value from the sorted_tasks's first record. But we don't have any
`planned_date_begin` in any task except the one we added in the Gantt view.

So the first task which is `path_last_task` will not execute further because of this,
https://github.com/odoo/enterprise/blob/c9a753fce85dfb09dc413831993e63c7736d258a/project_enterprise/models/project_task.py#L1529-L1531

This leads to the above traceback when we compare the value of
the `path_last_task` from the `total_time` with the current task.
https://github.com/odoo/enterprise/blob/c9a753fce85dfb09dc413831993e63c7736d258a/project_enterprise/models/project_task.py#L1544

We can resolve this issue by taking the `path_last_task` if only
the task contains `planned_date_begin`.

sentry-6327461743

closes odoo/enterprise#83869

X-original-commit: 2544d5dc553352d9d208c5119354be6d980f7df2
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Altaf Shaik (alsh) <alsh@odoo.com>