[FIX] documents_account: fix server actions
When you apply a journal entry action (e.g. Create Vendor Bill) to
several documents at once, you should be redirected to the journal list
view. Since version 18.0, you're redirected to the form view of the last
entry you've just created.

The corresponding actions have several sub-actions and cannot be
launched on multiple records at the same time. The selected documents are
processed individually and the last sub-action returns systematically a
form view.

To fix this, we inject a context key containing the ids of the
selection.Then the last sub-action uses it to determine the appropriate
view(s) to return.

task-4606751

closes odoo/enterprise#83007

X-original-commit: 89cfbf6daefb2b20c19b3a83dd7a0fd52f170356
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Marc Lopes (loma) <loma@odoo.com>