[FIX] appointment: disable booking action helper in pos
Purpose
=========
An unnecessary action helper was present in the pos Gantt view.

Technical
=========
https://github.com/odoo/enterprise/pull/70475/commits/427697ac326a7dc41b7e5a43634188b6d1650b69#diff-425f25c619962deff7d699cd67856a5081924b21d1f8f57190abcd719f1da464R26
Here we were using the active_model to prepare showNoContentHelp for the action
helper when we go into the Gantt from appointment but this context is also used
in the ```pos_restaurant_appointment```.
Added our own context for the custom action helper.

Task-4607266

closes odoo/enterprise#80464

X-original-commit: 61dc3311be6b48e5f4558ef4e4f6668cd6b42fe8
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>