[FIX] knowledge: don't show shared checkbox for embedded views
Steps to reproduce:

1. Add a embedded view in Knowledge, like a kanban view.
2. Add a new favorite filter.
3. The shared checkbox is shown.

To address this issue, we need to use the `.context` since in owl the
`._context` is giving inconsistent results, this due to how the rendering
is being done, we sometimes will try to access the context before it is
actually set, and we'll get a null value.

With the `.context` we are sure that the context is set and we can access
the `knowledgeEmbeddedViewId` properly.

opw

closes odoo/enterprise#80404

X-original-commit: 3a1926e959bd2e575c915b089e3352e6c8412005
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>