[FIX] l10n_in_reports: partial revert of https://github.com/odoo/enterprise/commit/344e09e103636d4134e43d806c96ce4c4347a0ef
In this commit we are partially reverting the company_domain changes
because using `child_of`, is not the best option here. Due to it
risks the using the credentials of the parent company for the child
company records where a child can have an independent credentials

The change was introduced because the in India, we can have same
PAN entity with multi-company or branches where we wanted to make
sure, that any child company does not have the record in that
case. But described above as said it is not the best option
to do.

closes odoo/enterprise#80328

Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>